dsv-controle-de-materiais-nti Corrigido erro em equipamentos e corrigo erro na hora de inserir e alterar administradores, não estava com md5

+++ intranet/excluir/exc_equipamento.php
@@ -1,12 +1,24 @@
 <?php
 
+include '../../conexao.php';
 $id = $_GET['cod'];
 $tipo = $_GET['id'];
 
-$sql = "DELETE FROM tb_equipamento WHERE cod_equipamento = $id";
-if ($pdo->query($sql)){
-    header("location: intranet.php?url=equip&msg=exc&id=$tipo");
-    exit;
+$verifica = "SELECT * FROM tb_equipamento WHERE cod_equipamento = $id";
+$result = $pdo->query($verifica);
+foreach ($result as $key) {
+    $res = $key['cod_equipamento'];
+}
+
+if (!empty($res)) {
+    header("location: intranet.php?url=equip&msg=erro&id=$tipo");
+} else {
+
+    $sql = "DELETE FROM tb_equipamento WHERE cod_equipamento = $id";
+    if ($pdo->query($sql)) {
+        header("location: intranet.php?url=equip&msg=exc&id=$tipo");
+        exit;
+    }
 }
